add SequentialAccess support to `RecordReaderBuilder`
