+++ Source/LinqToDB/Data/RecordReaderBuilder.cs
@@ -294,6 +294,9 @@ namespace LinqToDB.Data
 
 			var lambda = Expression.Lambda<Func<IDataReader,T>>(BuildBlock(expr), DataReaderParam);
 
+			if (Common.Configuration.OptimizeForSequentialAccess)
+				lambda = (Expression<Func<IDataReader, T>>)SequentialAccessHelper.OptimizeMapingExpressionForSequentialAccess(lambda, Reader.FieldCount, true);
+
 			return lambda.Compile();
 		}
